Hidden Rewards
Updated the function that collects all hidden rewards (still incomplete).
The function still needs to make sure the hidden rewards are collectable, I guess. For example, a pothole 2x2 isn't normally collectable if no big roads are on the city, but the collectAllReward function can still see it in the game files and collect it, which may be a problem